Feat(id): 迁移API边界ID类型

+++ sandwish-admin-api/src/main/java/com/github/thundax/modules/sys/controller/LogController.java
@@ -69,7 +69,7 @@ public class LogController {
     }
 
     private LogResponse toResponse(Log log) {
-        User user = userService.getById(EntityIdCodec.toDomain(log.getUserId()));
+        User user = userService.getById(EntityIdCodec.toDomain(Long.valueOf(log.getUserId())));
         Department department =
                 user == null ? null : departmentService.getById(EntityIdCodec.toDomain(user.getDepartmentId()));
         return LogInterfaceAssembler.toResponse(
